Added methods for applying the default 'checked' state to array checkboxes.
The normal old('field', $default) doesn't seem to work with checkboxes. The old() value isn't used properly if validation fails when the default value is true because it always uses the default of true. These methods just check to see if there are ANY old() input values, and if there are, then it will use the old() input, otherwise, it will use the given default.